refactor: service -> service 호출 방지 (프로젝트 컨벤션)
- StockPriceService -> StockPriceDataVO로 변경하여 비즈니스 로직 분리
